Use our own Nginx configs
This moves the global nginx config to a local config folder and switch
Procfile.dev to use that. I moved the minimal configs over for now and
we can keep adding new ones when necessary.

Because Nginx doesn't support variables in config well due to
performance reasons [1]. I had to use PHP to render proper project root
path in the pickee-wp conf. This is similar to how Heroku does it.

[1]. http://nginx.org/en/docs/faq/variables_in_config.html