Kill php-fpm processes when Ctrl+C exiting `foreman start -f Procfile.dev`
When running the current php-fpm processes via the script
(script/start_local_php-fpm.sh), Ctrl+C does not kill the php-fpm
processes, even the script process itself. Perhaps it's related to
bash's implementation of WCE (wait and cooperative exit).

Wrapping the php-fpm command in a subshell, trapping the INT signal, and
sending it to the subshell process manually solves it. We may explore
more sophisticated solution like Heroku's buildpack, but this seems to
be a quick fix.

https://unix.stackexchange.com/questions/230421/unable-to-stop-a-bash-script-with-ctrlc
https://github.com/heroku/heroku-buildpack-php/blob/06a81e71d94115668de3f202547fd9d517dc76ca/bin/heroku-php-nginx